Update product image and delete old image

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const Product = require("../models/product");
 const mongodb = require("mongodb");
 const { validationResult } = require("express-validator");
+const fileHelper = require('../util/file')
 exports.getAddProduct = (req, res, next) => {
 	// if(!req.session.isLoggedIn){
 	// 	return res.redirect('/login')
@@ -157,6 +158,7 @@ exports.postEditProduct = async (req, res, next) => {
 		product.price = updatedPrice;
 		product.description = updatedDesc;
 		if(image){
+			fileHelper.deleteFile(product.imageUrl)
 			product.imageUrl = image.path;
 		}
 		product.save();
@@ -187,6 +189,11 @@ exports.getProducts = async (req, res, next) => {
 
 exports.postDeleteProduct = async (req, res, next) => {
 	const prodId = req.body.productId;
+	const product = await Product.findById(prodId)
+	if(!product){
+		return next(new Error("Product not found"))
+	}
+	fileHelper.deleteFile(product.imageUrl)
 	try {
 		console.log("user =>", req.user._id);
 		await Product.deleteOne({ _id: prodId, userId: req.user._id });
